requirements.txt
aiohttp
async-timeout
attrs
chardet
discord.py
idna
multidict
py-getch
six
websockets
wrapt
yarl
